SSE-10 SSE Support for broadcasting Answer Summaries
- Refactors the Answer and AnswerKey out from the Puzzle.
