+++ src/providers/puzzle/answer.ts
@@ -1,0 +1,12 @@
+export class Answer {
+  id: number;
+  answerKey: AnswerKey;
+  answer: string;
+}
+
+/**
+ * Possible answer keys.
+ */
+export enum AnswerKey {
+  A, B, C, D, E
+}
