+++ src/providers/puzzle/puzzle.ts
@@ -1,6 +1,8 @@
 /**
  * Created by jett on 10/24/17.
  */
+import {Answer, AnswerKey} from "./answer";
+
 /**
  * Defines properties of a Puzzle.
  * Many-to-one per Location
@@ -16,16 +18,3 @@ export class Puzzle {
   points: number;
 }
 
-export class Answer {
-  id: number;
-  // puzzleId: number;
-  answerKey: AnswerKey;
-  answer: string;
-}
-
-/**
- * Possible answer keys.
- */
-export enum AnswerKey {
-  A, B, C, D, E
-}
